gift subscriptions: include sub event in gift event to fix dependent behaviour
If the subscription is considered a duplicate, that should also stop the gifter to get tokens a second time.
This requires the subscription event to be part of the gift event and not be a separate independent event,
otherwise the outcome of processing the subscription could not influence the processing of the gift.