fix some log messages being dropped when aborting program
the Abort handler for the ProcessExit and CancelKeyPress event was waiting for the mode to finish,
but failed to account for the fact that the process immediately terminates after that was done,
which in terms does not properly dispose stuff like the logger factory.
The latter is required for any remaining log messages to be flushed out.